pyproject.toml
[tool.poetry]
name = "myrecipes"
version = "0.1.0"
description = ""
authors = ["Sebastian <se.hegardt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
recipe-scrapers = "^14.12.0"
requests = "^2.28.1"
click = "^8.1.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
